♻️ 메인, 편지 작성 NavBar 리팩토링 및 모듈화
