+++ src/components/common/TopNavigation/atoms/NavBack.tsx
@@ -1,23 +1,20 @@
 import Image from 'next/image';
-import { useRouter } from 'next/router';
+import Link from 'next/link';
 import { ReactElement } from 'react';
 import * as S from './styled';
 
 const NavBack = (): ReactElement => {
-  const router = useRouter();
   return (
-    <S.NavButton
-      onClick={() => {
-        router.back();
-      }}
-    >
-      <Image
-        src='/icons/icon__nav-back.svg'
-        alt='뒤로 가기'
-        width={24}
-        height={24}
-      />
-    </S.NavButton>
+    <Link href='/'>
+      <S.NavItemContainer width={24} height={24}>
+        <Image
+          src='/icons/icon__nav-back.svg'
+          alt='뒤로 가기'
+          width={24}
+          height={24}
+        />
+      </S.NavItemContainer>
+    </Link>
   );
 };
