+++ src/components/common/TopNavigation/atoms/index.ts
@@ -1,4 +1,6 @@
+import NavLogo from './NavLogo';
+import NavBell from './NavBell';
 import NavBack from './NavBack';
 import NavCancel from './NavCancel';
 
-export { NavBack, NavCancel };
+export { NavLogo, NavBell, NavBack, NavCancel };
