--- src/components/common/gnb/index.ts
@@ -1,4 +1,0 @@
-import NavBell from './NavBell';
-import NavLogo from './NavLogo';
-
-export { NavBell, NavLogo };
